refactor: migrate all financial amounts to BigDecimal, add rejectedAt field, harden validation
- Migrate double → BigDecimal across all domain entities, DTOs, service layer, and demo data
  (ExpenseItem, ExpenseReport, ExpenseItemCreateRequest, ExpenseItemResponse,
   ExpenseReportResponse, ExpenseReportListItemResponse, StatsResponse,
   ExpenseReportActivityItem, PolicyEngine, DemoDataService, ExpenseReportService)
- Fix approvedAt/rejectedAt bug: rejection now correctly sets rejectedAt instead of approvedAt
- Add @Email validation to LoginRequest
- Add DataIntegrityViolationException handler (409 Conflict) to GlobalExceptionHandler
- Add DB indexes on AuditLog (report_id, createdAt)
- Add @Column constraints (unique/nullable) on User entity fields
- Add @Index on ExpenseItem.expense_report_id
- Increase approvalComment max length to 2000, description to 500
- Add @Positive + @DecimalMax("999999.99") on ExpenseItemCreateRequest.amount
- Update recentActivity JPQL to include rejectedAt in coalesce sort
- Fix BigDecimal sort comparisons in searchReports (was using Double.compare)

+++ src/main/java/com/example/demo/dto/ExpenseReportListItemResponse.java
@@ -4,6 +4,7 @@ package com.example.demo.dto;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -12,15 +13,15 @@ public class ExpenseReportListItemResponse {
 
     private Long id;
     private String title;
-    private double totalAmount;
+    private BigDecimal totalAmount;
     private String status;
     private String destination;
     private LocalDate departureDate;
     private LocalDate returnDate;
 
     // Per-diem
-    private double perDiemAmount;
-    private double perDiemRate;
+    private BigDecimal perDiemAmount;
+    private BigDecimal perDiemRate;
     private int perDiemDays;
 
     // Demo policy flags (for list/search UI)
